Added inline_svg image format to render svg's inline
Using format specifier `inline_svg` the svg image produced by plantuml
will be embedded into html page.
This can be usefull to have hyperlinks from elements of the diagram to
wiki pages or other external links.
The disadvantage is a biggest page and a slightly lower performance of
redmine in rendering the page.